requirements.txt
flask==2.3.3
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
textblob==0.17.1
PyPDF2==3.0.1
pdfplumber==0.9.0
pytesseract==0.3.10
Pillow==10.0.1
python-docx==0.8.11
plotly==5.17.0
seaborn==0.12.2
matplotlib==3.7.2
werkzeug==2.3.7
